+++ src/test/kotlin/com/github/hechtcarmel/jetbrainsindexmcpplugin/tools/ToolsUnitTest.kt
@@ -194,6 +194,7 @@ class ToolsUnitTest : TestCase() {
         assertNotNull("Should have column property", properties?.get(ParamNames.COLUMN))
         assertNotNull("Should have className property", properties?.get(ParamNames.CLASS_NAME))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
     }
 
     fun testCallHierarchyToolSchema() {
@@ -212,6 +213,7 @@ class ToolsUnitTest : TestCase() {
         assertNotNull("Should have language property", properties?.get(ParamNames.LANGUAGE))
         assertNotNull("Should have symbol property", properties?.get(ParamNames.SYMBOL))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
     }
 
     fun testFindImplementationsToolSchema() {
@@ -232,6 +234,7 @@ class ToolsUnitTest : TestCase() {
         assertNotNull("Should have language property", properties?.get(ParamNames.LANGUAGE))
         assertNotNull("Should have symbol property", properties?.get(ParamNames.SYMBOL))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
         assertNotNull("Should have cursor property", properties?.get("cursor"))
         assertNotNull("Should have pageSize property", properties?.get("pageSize"))
 
@@ -245,6 +248,7 @@ class ToolsUnitTest : TestCase() {
 
         assertNotNull("Should have query property", properties?.get(ParamNames.QUERY))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
     }
 
     fun testFindFileToolSchemaUsesScope() {
@@ -253,6 +257,7 @@ class ToolsUnitTest : TestCase() {
 
         assertNotNull("Should have query property", properties?.get(ParamNames.QUERY))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
     }
 
     fun testFindSymbolToolSchemaUsesScope() {
@@ -261,6 +266,7 @@ class ToolsUnitTest : TestCase() {
 
         assertNotNull("Should have query property", properties?.get(ParamNames.QUERY))
         assertHasScopeAndNoLegacyFilters(properties)
+        assertNotNull("Should have includeGenerated property", properties?.get(ParamNames.INCLUDE_GENERATED))
     }
 
     fun testGetDiagnosticsToolSchema() {